OTA testing 4G, configuration parameters for fairRR scheduler and common parameters for both fairRR and default

+++ openair2/F1AP/f1ap_du_task.c
@@ -88,7 +88,7 @@ void du_task_handle_sctp_association_resp(instance_t instance, sctp_new_associat
   if (sctp_new_association_resp->sctp_state != SCTP_STATE_ESTABLISHED) {
     LOG_E(F1AP, "Received unsuccessful result for SCTP association (%u), instance %d, cnx_id %u\n",
               sctp_new_association_resp->sctp_state,
-              instance,
+              (int)instance,
               sctp_new_association_resp->ulp_cnx_id);
 
       //f1ap_handle_setup_message(instance, sctp_new_association_resp->sctp_state == SCTP_STATE_SHUTDOWN);
